vending-machine-controller/requirements.txt
Flask>=2.0.0
pyserial>=3.5
requests>=2.28.0
